+++ .eslintrc.json
@@ -14,6 +14,7 @@
     }],
     "object-curly-spacing": ["error", "always"],
     "space-before-blocks": ["error", { "functions": "always", "keywords": "always", "classes": "always" }],
-    "space-before-function-paren": ["error", {"anonymous": "never", "named": "never", "asyncArrow": "always"}]
+    "space-before-function-paren": ["error", {"anonymous": "never", "named": "never", "asyncArrow": "always"}],
+    "no-multi-spaces": ["error"]
   }
 }
