+++ pages/index.module.scss
@@ -26,22 +26,3 @@
     margin: 24px 0 18px;
   }
 }
-
-.blogPosts {
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-  gap: 16px;
-  margin: 0;
-  padding: 0;
-  list-style-type: none;
-
-  @media (max-width: $bp-md) {
-    gap: 12px;
-    grid-template-columns: 1fr;
-  }
-}
-
-.blogPostsItem {
-  display: flex;
-  width: 100%;
-}
